redis_poolex v0.0.6 RedisPoolex

Application for running connection pool and redis connection inside.

Example:

alias RedisPoolex, as: Redis

  Redis.query(["SET", "key1", "value1"]) => "OK"
  Redis.query(["GET", "key1"]) => "value1"
  Redis.query(["GET", "key2"]) => :undefined

Functions

query(args)

query sends commands directly to Redis

query_pipe(args)

query_pipe sends multiple commands as batch directly to Redis.

start(type, args)

Callback implementation for c:Application.start/2.