redis_poolex v0.0.4 RedisPoolex.Supervisor

Redis connection pool supervisor to handle connections via pool and reduce the number of opened connections via GenServer.

Functions

init(list)

Callback implementation for c::supervisor.init/1.

q(args)

Making query via connection pool using %{command: command, params: params} pattern.

start_link()